Creating a nglpy_cuda compatible API for nglpy.
